tests: postprocessor
